+++ tests/scripts/wizard_monitors/branch-br_table.mm
@@ -30,7 +30,7 @@ wasm::br_if:before {
 wasm::br_table:before {
   // which branch was taken?
   var index: i32;
-  index = arg0 < (num_targets - 1) ? targets[arg0] as i32 : default_target as i32;
+  index = arg0 < (num_targets - 1) ? targets[arg0 as u32] as i32 : default_target as i32;
 
   // count stores an array of counters
   count[(fid, pc, index)]++;
